admin tausif

+++ src/main/java/com/aurionpro/loan/repository/EmiRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.aurionpro.loan.entity.Emi;
 
-public interface EmiRepository extends JpaRepository<Emi, Integer> {
+public interface EmiRepository extends JpaRepository<Emi, Long> {
 
 }
